requirements.txt
Flask==2.3.3
opencv-python-headless==4.8.0.76
Pillow==10.0.0
numpy==1.24.3
gunicorn==21.2.0
